refactor: refactor span to glyph
