+++ .vscode/settings.json
@@ -51,6 +51,7 @@
         "Lerp",
         "linebreak",
         "linebreaker",
+        "linebreaking",
         "Liti",
         "localforage",
         "LRTBV",
