+++ packages/engine-render/src/components/docs/layout/block/paragraph/layout.ts
@@ -14,37 +14,22 @@
  * limitations under the License.
  */
 
-import type { IBullet, ICustomBlock, IDrawing, IDrawings, Nullable } from '@univerjs/core';
-import {
-    DataStreamTreeTokenType,
-    HorizontalAlign,
-    PositionedObjectLayoutType,
-} from '@univerjs/core';
-
+import { HorizontalAlign } from '@univerjs/core';
 import type {
-    IDocumentSkeletonBullet,
-    IDocumentSkeletonDrawing,
     IDocumentSkeletonLine,
     IDocumentSkeletonPage,
     ISkeletonResourceReference,
 } from '../../../../../basics/i-document-skeleton-cached';
-import { BreakType } from '../../../../../basics/i-document-skeleton-cached';
-import type { IParagraphConfig, ISectionBreakConfig } from '../../../../../basics/interfaces';
-
-import { createSkeletonPage } from '../../model/page';
-import { setColumnFullState } from '../../model/section';
+import type { ISectionBreakConfig } from '../../../../../basics/interfaces';
 import {
     clearFontCreateConfigCache,
-    getLastNotFullColumnInfo,
     getSpanGroupWidth,
     lineIterator,
 } from '../../tools';
 import type { DataStreamTreeNode } from '../../../view-model/data-stream-tree-node';
 import type { DocumentViewModel } from '../../../view-model/document-view-model';
-import { dealWithBullet } from './bullet';
-import { dealWidthInlineDrawing } from './inline-drawing';
-import { layoutParagraph } from './layout-ruler';
 import { shaping } from './shaping';
+import { lineBreaking } from './linebreaking';
 
 export function dealWidthParagraph(
     bodyModel: DocumentViewModel,
@@ -53,99 +38,13 @@ export function dealWidthParagraph(
     sectionBreakConfig: ISectionBreakConfig,
     skeletonResourceReference: ISkeletonResourceReference
 ): IDocumentSkeletonPage[] {
-    const {
-        pageSize: _pageSize = {
-            width: Number.POSITIVE_INFINITY,
-            height: Number.POSITIVE_INFINITY,
-        },
-
-        marginRight: _marginRight = 0,
-        marginLeft: _marginLeft = 0,
-
-        lists,
-        drawings = {},
-        localeService,
-    } = sectionBreakConfig;
-
-    const { endIndex, startIndex, content = '', blocks = [] } = paragraphNode;
-
-    const paragraph = bodyModel.getParagraph(endIndex) || { startIndex: 0 };
-
-    const { paragraphStyle = {}, bullet } = paragraph;
-
-    // const paragraphAffectSkeDrawings = _changeDrawingToSkeletonFormat(drawingIds, drawings);
-
-    // curPage.skeDrawings = paragraphAffectSkeDrawings;
-
-    const { skeHeaders, skeFooters, skeListLevel, drawingAnchor } = skeletonResourceReference;
-
-    // const pageWidth = pageSize.w || Infinity;
-    // const pageHeight = pageSize.h || Infinity;
-
-    // const headerAndFooterAffectSkeDrawings = new Map([
-    //     ...(skeHeaders?.get(curPage.headerId)?.get(pageWidth)?.skeDrawings || []),
-    //     ...(skeFooters?.get(curPage.footerId)?.get(pageWidth)?.skeDrawings || []),
-    // ]);
-
-    // curPage.skeDrawings = affectSkeDrawings; // 加入本段落的对象
-
-    // const pages = [curPage];
-    // let lastPage = curPage;
-
-    // if (pageSize.h === Infinity) {
-    //     // 无需分页的情况，适合表格单元格或者类似notion的block布局
-    // }
-
-    // if (curPage.sections.length > 0) {
-    //     // 处理接上一页的情况
-    //     let columnHeight = _getColumnHeight(curPage);
-    //     curPage.skeDrawings;
-
-    //     return [];
-    // }
-
-    const paragraphAffectSkeDrawings: Map<string, IDocumentSkeletonDrawing> = new Map();
-
-    const paragraphConfig: IParagraphConfig = {
-        paragraphIndex: endIndex,
-        paragraphStyle,
-        paragraphAffectSkeDrawings,
-        skeHeaders,
-        skeFooters,
-        drawingAnchor,
-    };
-
-    const listLevelAncestors = _getListLevelAncestors(bullet, skeListLevel); // 取得列表所有 level 的缓存
-    const bulletSkeleton = dealWithBullet(bullet, lists, listLevelAncestors, localeService); // 生成 bullet
-
-    _updateListLevelAncestors(bullet, bulletSkeleton, skeListLevel); // 更新最新的 level 缓存列表
-
-    paragraphConfig.bulletSkeleton = bulletSkeleton;
-
-    const { horizontalAlign = HorizontalAlign.UNSPECIFIED } = paragraphStyle;
-
-    let allPages = [curPage];
-
-    const customBlockCache = new Map<number, Nullable<ICustomBlock>>();
-
-    for (let i = 0, len = blocks.length; i < len; i++) {
-        const charIndex = blocks[i];
-        const customBlock = bodyModel.getCustomBlock(charIndex);
-        if (customBlock == null) {
-            continue;
-        }
-
-        customBlockCache.set(charIndex, customBlock);
-        const blockId = customBlock.blockId;
-        const drawingOrigin = drawings[blockId];
-        if (drawingOrigin.layoutType !== PositionedObjectLayoutType.INLINE) {
-            paragraphAffectSkeDrawings.set(blockId, _getDrawingSkeletonFormat(drawingOrigin));
-        }
-    }
-
     clearFontCreateConfigCache();
 
     // Step 1: Text Shaping.
+    const { endIndex, content = '' } = paragraphNode;
+    const paragraph = bodyModel.getParagraph(endIndex) || { startIndex: 0 };
+    const { paragraphStyle = {} } = paragraph;
+
     const shapedTextList = shaping(
         content,
         bodyModel,
@@ -155,165 +54,24 @@ export function dealWidthParagraph(
     );
 
     // Step 2: Line Breaking.
-    let paragraphStart = true;
-    let charIndex = startIndex;
-    for (const { text, glyphs } of shapedTextList) {
-        charIndex += text.length;
-        const pushPending = () => {
-            if (glyphs.length === 0) {
-                return;
-            }
-
-            allPages = layoutParagraph(
-                glyphs,
-                allPages,
-                sectionBreakConfig,
-                paragraphConfig,
-                paragraphStart
-            );
-
-            paragraphStart = false;
-        };
-
-        if (text.endsWith(DataStreamTreeTokenType.CUSTOM_BLOCK)) {
-            let customBlock = customBlockCache.get(charIndex);
-            if (customBlock == null) {
-                customBlock = bodyModel.getCustomBlock(charIndex);
-            }
-            if (customBlock != null) {
-                const blockId = customBlock.blockId;
-                const drawingOrigin = drawings[blockId];
-                if (drawingOrigin.layoutType === PositionedObjectLayoutType.INLINE) {
-                    allPages = dealWidthInlineDrawing(
-                        drawingOrigin,
-                        sectionBreakConfig,
-                        allPages,
-                        paragraphConfig,
-                        localeService
-                    );
-                }
-            }
-            pushPending();
-            continue;
-        } else if (text.endsWith(DataStreamTreeTokenType.PAGE_BREAK)) {
-            pushPending();
-            allPages.push(
-                createSkeletonPage(
-                    sectionBreakConfig,
-                    skeletonResourceReference,
-                    _getNextPageNumber(allPages[allPages.length - 1]),
-                    BreakType.PAGE
-                )
-            );
-            paragraphAffectSkeDrawings.clear();
-            continue;
-        } else if (text.endsWith(DataStreamTreeTokenType.COLUMN_BREAK)) {
-            pushPending();
-            // 换列标识，还在同一个节内
-            const lastPage = allPages[allPages.length - 1];
-            const columnInfo = getLastNotFullColumnInfo(lastPage);
-
-            if (columnInfo && !columnInfo.isLast) {
-                setColumnFullState(columnInfo.column, true);
-            } else {
-                allPages.push(
-                    createSkeletonPage(
-                        sectionBreakConfig,
-                        skeletonResourceReference,
-                        _getNextPageNumber(lastPage),
-                        BreakType.COLUMN
-                    )
-                );
-            }
-            continue;
-        }
-
-        pushPending();
-    }
+    const allPages = lineBreaking(
+        shapedTextList,
+        curPage,
+        bodyModel,
+        paragraphNode,
+        sectionBreakConfig,
+        skeletonResourceReference
+    );
 
     // Step 3: Line Adjustment.
     // Handle horizontal align: left\center\right\justified.
+
+    const { horizontalAlign = HorizontalAlign.UNSPECIFIED } = paragraphStyle;
     lineIterator(allPages, (line) => {
         horizontalAlignHandler(line, horizontalAlign);
     });
 
     return allPages;
-}
-
-function _getListLevelAncestors(
-    bullet?: IBullet,
-    listLevel?: Map<string, IDocumentSkeletonBullet[]>
-): Array<Nullable<IDocumentSkeletonBullet>> | undefined {
-    if (!bullet || !listLevel) {
-        return;
-    }
-
-    const { listId, nestingLevel } = bullet;
-
-    const sameList = listLevel?.get(listId);
-
-    let level = nestingLevel;
-
-    if (level < 0) {
-        level = 0;
-    }
-
-    const listLevelAncestors: Array<Nullable<IDocumentSkeletonBullet>> = [];
-
-    for (let i = level; i >= 0; i--) {
-        const bs = sameList?.[i];
-        listLevelAncestors[i] = bs || null;
-    }
-
-    return listLevelAncestors;
-}
-
-function _updateListLevelAncestors(
-    bullet?: IBullet,
-    bulletSkeleton?: IDocumentSkeletonBullet,
-    listLevel?: Map<string, IDocumentSkeletonBullet[]>
-) {
-    if (!bullet || !bulletSkeleton) {
-        return;
-    }
-
-    const { listId, nestingLevel } = bullet;
-
-    const cacheItem: IDocumentSkeletonBullet[] = [...(listLevel?.get(listId) || [])];
-
-    // [[nestingLevel, bulletSkeleton]];
-
-    cacheItem[nestingLevel] = bulletSkeleton;
-
-    cacheItem.splice(nestingLevel + 1); // 文档自上而下渲染，如果一个level被更新，则它以下的level数据的startIndex就要重置
-
-    listLevel?.set(listId, cacheItem);
-}
-
-function _changeDrawingToSkeletonFormat(
-    drawingIds: string[],
-    drawings: IDrawings
-): Map<string, IDocumentSkeletonDrawing> {
-    const skeDrawings: Map<string, IDocumentSkeletonDrawing> = new Map();
-    drawingIds.forEach((objectId) => {
-        const drawingOrigin = drawings[objectId];
-        drawingOrigin && skeDrawings.set(objectId, _getDrawingSkeletonFormat(drawingOrigin));
-    });
-    return skeDrawings;
-}
-
-function _getDrawingSkeletonFormat(drawingOrigin: IDrawing) {
-    const objectId = drawingOrigin.objectId;
-    return {
-        objectId,
-        aLeft: 0,
-        aTop: 0,
-        width: 0,
-        height: 0,
-        angle: 0,
-        initialState: false,
-        drawingOrigin,
-    };
 }
 
 function _checkAndPush(pages: IDocumentSkeletonPage[], currentPages: IDocumentSkeletonPage[]) {
@@ -339,10 +97,6 @@ function _checkAndPush(pages: IDocumentSkeletonPage[], currentPages: IDocumentSk
 //         ...paragraphAffectSkeDrawings,
 //     ]);
 // }
-
-function _getNextPageNumber(lastPage: IDocumentSkeletonPage) {
-    return lastPage.pageNumber + 1;
-}
 
 /**
  * todo: @author DR-univer
